--- src/app/users/[userId]/inbox/_handlers/create.ts
@@ -1,8 +1,0 @@
-import { inboxCreateService } from "@/server/service/inbox";
-
-import type { InboxFunction } from "./types";
-
-export const create: InboxFunction = async (activity, actorUser) => {
-  await inboxCreateService.handle(activity, actorUser);
-  return { status: 200, message: "完了: ノート作成" };
-};
