--- src/app/users/[userId]/inbox/_handlers/types.ts
@@ -1,9 +1,0 @@
-import type { User } from "@prisma/client";
-
-export type InboxFunction = (
-  activity: unknown,
-  actorUser: User,
-) => Promise<{
-  status: number;
-  message: string;
-}>;
