+++ src/server/service/inbox/index.ts
@@ -1,6 +1,1 @@
-export * as inboxAcceptService from "./accept";
-export * as inboxCreateService from "./create";
-export * as inboxDeleteService from "./delete";
-export * as inboxFollowService from "./follow";
-export * as inboxLikeService from "./like";
-export * as inboxUndoService from "./undo";
+export * as inboxService from "./inbox";
